+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Switched from Black to Ruff for code formatting
 * Updating README
 * Rewrite of set dates script
+* Rewrite of set transfer options script
 
 ## [0.1.0] - 2023-11-10